requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
SQLAlchemy==2.0.23
Werkzeug==2.3.7
email-validator==2.1.0
pycryptodome==3.19.0
python-dotenv==1.0.0
WTForms==3.1.1
imaplib2==2.65.0
pytest==7.4.3
gunicorn==21.2.0 